handshake: support record size limit extension
RFC 8449 [1] Section 4 defines the record_size_limit TLS extension, which
allows peers to negotiate a maximum plaintext record size during the
TLS handshake. The value must be between 64 bytes and 16,384 bytes (2^14).
If a TLS endpoint receives a record larger than its advertised limit, it
must send a fatal record_overflow alert.

This patch adds support for extracting the record size limit value from
the gnutls session and passing it to the kernel. This is to be used by
the kernel at the tls layer limit outgoing records to the maximum
record size limit as specified by the endpoint.

[1] https://www.rfc-editor.org/rfc/rfc8449#section-4

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>